upload-service/requirements.txt
aio_pika==9.5.8
beautifulsoup4==4.14.2
chardet==5.2.0
pydantic==2.12.5
pydantic_settings==2.12.0
pypandoc==1.16.2
pypdf==6.4.0
python_docx==1.2.0
